Remove LiveCharts dependency
No LiveCharts controls remain, so remove the dependency and remove
any remaining references.

This avoids bundling extra DLLs that are not needed.

LiveCharts was an example of how third party dependencies could be
included in a custom analysis and how they could be used to include
UI elements not exposed by AP Suite or in the standard Microsoft
control library.